reorder imports, rename partialSyncHelper file, and clarify comment
